website-root-shroomworks/requirements.txt
djangorestframework==3.5.3
django-extensions==1.7.5
